+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 
+### Fixed
+- Error when no file shares are available #12
+
 ## 1.3.3 - 2025-01-03
 ### Added
 - Federation shares